Added barrelling for media controls
